subpackages also need some docs
