+++ sql/sql.go
@@ -1,10 +1,5 @@
 package sql
 
-type SelectStmt struct {
-	Columns []string
-	Table   string
-}
-
 type PrimaryKey int
 
 const (
@@ -13,6 +8,7 @@ const (
 	PKDesc
 )
 
+// Type of a column, as found by CreateTableStmt
 type ColumnDef struct {
 	Name          string
 	Type          string
@@ -26,7 +22,26 @@ type ColumnDef struct {
 	// foreign key
 }
 
+// A `SELECT` statement
+type SelectStmt struct {
+	Columns []string
+	Table   string
+}
+
+// A `CREATE TABLE` statement
 type CreateTableStmt struct {
 	Table   string
 	Columns []ColumnDef
+}
+
+// Parse is the main function. It will return either an error or a *Stmt
+// struct.
+func Parse(sql string) (interface{}, error) {
+	ts, err := tokenize(sql)
+	if err != nil {
+		return nil, err
+	}
+	l := &Lexer{tokens: ts}
+	yyParse(l)
+	return l.result, l.err
 }
